// src/modules/products/dto/list-products/list-products-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

class PaginationMetadataDto {
  @ApiProperty({ description: 'Total products found by call' })
  total: number;

  @ApiProperty({ description: 'Products on current page' })
  count: number;

  @ApiProperty({ description: 'Product pagination request' })
  per_page: number;

  @ApiProperty({ description: 'Current page' })
  current_page: number;

  @ApiProperty({ description: 'Total pages generated by request' })
  total_pages: number;
}

export class ListProductsResponseDto {
  @ApiProperty({ description: 'The list of products', type: [Object] })
  data: any[];

  @ApiProperty({ description: 'Pagination metadata', type: PaginationMetadataDto })
  pagination: PaginationMetadataDto;
}